requirements-full.txt
# Production Dependencies for FastAPI API
# Install with: pip install -r requirements.txt

# Core ML Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
imbalanced-learn>=0.11.0  # If you're handling imbalanced data

# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.25.0
pydantic==2.10.3
python-multipart==0.0.6

# Logging and Monitoring
python-json-logger==2.0.7

# Core Data Science Libraries 
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

#MLOps and Experiment Tracking
mlflow>=2.8.0,<3.0.0
dvc>=3.30.0,<4.0.0
dvc-s3>=3.0.0 
boto3>=1.28.0,<2.0.0  # AWS SDK for Python (used by dvc-s3)

# Data Validation and Monitoring
evidently>=0.4.0,<1.0.0
deepchecks>=0.17.0,<1.0.0

# Synthetic Data (if needed)
sdv>=1.0.0,<2.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0  # For test coverage reports
pytest-mock>=3.11.0
coverage>=7.3.0


# Development tools
notebook>=7.0.0
jupyterlab>=4.0.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Notebook Execution
papermill>=2.4.0  # For automated notebook execution
nbconvert>=7.8.0  # For converting notebooks to other formats

# Configuration Management
pyyaml==6.0.1  # For config.yaml files
python-dotenv>=1.0.0  # For environment variables

# Utilities
tqdm>=4.66.0  # Progress bars
click>=8.1.0  # CLI tool (if you use command-line scripts)

# Data Format Support
openpyxl>=3.1.0  # Excel file support
xlrd>=2.0.0  # Excel file reading

# Model Serialization (scikit learn uses it)
joblib>=1.3.0 

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Project Template
cookiecutter-data-science

# Web Framework (if applicable)
fastapi>=0.100.0,<1.0.0